In-depth analysis of IC data cubes

gg_point This plot function is used to produce the scatter plots of Figure 8 and Supplementary Figure 9.

Usage

gg_point(
  title,
  grid_cell,
  ion1_thr,
  ion2_thr,
  thr,
  ion1_R,
  ion2_R,
  colors = c("#8E063B", "#023FA5"),
  save = FALSE,
  .X = NULL,
  .N = NULL,
  .species = NULL,
  .t = NULL
)

Arguments

title

Character string for the analyte ("MEX" or "MON") to be used.

grid_cell

Selected grid cell.

ion1_thr

Character string for ion in the enumerator of ion ratio of the map.

ion2_thr

Character string for ion in the denominator of ion ratio of the map.

thr

Numeric threshold value for filter selection.

ion1_R

A character string constituting the rare isotope ("13C").

ion2_R

A character string constituting the common isotope ("12C").

colors

Colour palette to identify matrix and inclusion (default = c("#8E063B", "#023FA5")).

save

Boolean whether to save the plot as an png.

.X

Variable for ion count rates (default = NULL).

.N

Variable for ion counts (default = NULL).

.species

Variable for species names (default = NULL).

.t

Variable for time increment (default = NULL).

Value

ggplot2::ggplot.