Manipulate ion count data cubes

grid_aggregate Aggregate ion count data cubes on a 2D grid. grid_select Select ion count data cubes from a grid. tune_grid_cell Allows calibration of the grid_cell size.

Usage

grid_aggregate(
  IC,
  plane,
  grid_cell = NULL,
  select_cell = NULL,
  species = NULL,
  title = character(1),
  name = character(1),
  corrected = FALSE,
  scalar = 40/256,
  save = FALSE
)

grid_select(
  IC,
  plane,
  grid_cell = NULL,
  select_cell = NULL,
  species = NULL,
  title = character(1),
  name = character(1),
  corrected = FALSE,
  scalar = 40/256,
  save = FALSE
)

tune_grid_cell(grid_expr, tune, mc.cores = 1)

Arguments

IC

list of Ion Count (IC) data cubes.

plane

Character string for the dimension over which to be aggregated (vectors are allowed).

grid_cell

Integer indicating size of cells in pixels, only exponent of base two allowed (default = NULL). If NULL then 2D raster images are created.

select_cell

Integer indicating a grid cell to be extracted.

species

Character string or vector with chemical species. Default is NULL and is extracted from the IC data cubes. When supplied it will be used to filter the list of IC data cubes.

title

Character string for file name.

name

Character string for sample name.

corrected

If TRUE then IC is corrected with point::cor_IC().

scalar

Numeric converting the pixels to metric dimension of measurement (default is the conversion used in this study).

save

Boolean indicating whether to save the file in the package internal data store (default = FALSE).

grid_expr

The function `grid_select` or `grid_aggregate` to be tuned.

tune

A numeric vector for grid_cell size tuning (following: 2 ^ x).

mc.cores

An integer value for multi-threading.

Value

A tibble::tibble containing the ion counts aggregated or extracted over the plane of choice.

Details

These functions are designed to work with data cubes as produced by the Matlab based program Look At NanoSims (LANS). They allow sub-setting as well as flattening of the cube. This is required for calculating high precision isotope estimates.